backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
pymupdf==1.25.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pandas==2.2.3
openpyxl==3.1.5
bcrypt==4.2.1
passlib[bcrypt]==1.7.4